Extract the Authz module (role to permission mapping) into Api/Authz
Sixth module of the #777 folder migration. Moves the generic role→permission
types into Jellyfin.Plugin.SSO_Auth.Api.Authz:

- PermissionGrant, PermissionRolePolicy, RolePrivilegeMapper -> Api/Authz/
  (namespace Api.Authz), import added at each call site.

Authz is a leaf (no Api-module dependency). OidcRoleExtractor is intentionally
left in the flat Api namespace: it is OIDC-flavoured claim extraction and will
move with the Oidc module, not here.

Pure mechanical move, no behavior change. Verified by the full build
(--warnaserror, both TFMs), the complete test suite (1577 both TFMs), the module
DAG conformance rule, and the ABI-floor build.

Refs #777
