+++ SSO-Auth/Api/Authz/RolePrivilegeMapper.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Jellyfin.Plugin.SSO_Auth.Config;
 
-namespace Jellyfin.Plugin.SSO_Auth.Api;
+namespace Jellyfin.Plugin.SSO_Auth.Api.Authz;
 
 /// <summary>
 /// Maps the roles carried by a verified login (OpenID or SAML) to the privileges they grant, according
