+++ SSO-Auth/Api/SamlAuthorizeStateBuilder.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using Jellyfin.Plugin.SSO_Auth.Api.Authz;
 using Jellyfin.Plugin.SSO_Auth.Config;
 
 #nullable enable
