+++ SSO-Auth/Api/SessionParameters.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Jellyfin.Plugin.SSO_Auth.Api.Authz;
 
 #nullable enable
